fix: 🐛 always make sure all bootstrapees are upt to date
New methods in bootstrap that figure out the execution order for the
projects to be bootstraps based on the dependency tree. Sometimes there
are circular dependencies and the method can generate an order with
repeats to make sure that every dependency is built and linked so they
are all fully up to date. Optimization for the future: re-implement to
link or not to link to not waste time.