fix: 🐛 bootstrap rewrite
Bootstrap, local, and addRemove rewritten from the ground up.
Dependencies are now built in two versions, a production version with
everything inside and the local version with zero local dependencies.
Local tarballs are stored outside the directory. On local add/remove and
in bootstrap, polyman looks for the dependencies and "tricks" yarn into
believing all dependencies are regular dependencies that do not require
each other. This ensures that the bootstrapping package is always the
most up to date possible verion, as there will never be conflicts
betweeen subdependencies and dependencies, cache and tmp conflicts, etc.