fix: 🐛 yarn cache directory found programatically
THIS was the bug. I had an environment variable setting the yarn cache
location that disappears in other project directories, making cache
clearing fail elsewhere but succeed during testing. Without cache clear,
most up to date versions of projects cannot be loaded because yarn will
refer to the cache instead of the tarball.