feat: 🎸 standardize built tarball dir
Standardized the built tarball director name generation. Tests for a
number of methods. Cleanup of utilities for more precise abort behavior
and removing unneccesary findRepository calls on the path to making poly
run without being in a repository

BREAKING CHANGE: 🧨 Build return changed to export whether it ran and the path the function
built to.